Split the TimeSeries definition normalization into two files
`normalizeTimeSeriesDefinitionImpl` gets its own header and source with
`NormalizeTimeSeriesDefinitionParams`, which now also carries the loading
mode and the restore flag; `normalizeTimeSeriesDefinition` only reads the
catalog and calls it. Generated inner engines are formatted without empty
parentheses, the same way the parser reads them, so a stored definition
stays unchanged when normalized again.
