+++ src/Storages/TimeSeries/tests/gtest_normalize_time_series_definition.cpp
@@ -1,4 +1,4 @@
-#include <Storages/TimeSeries/normalizeTimeSeriesDefinition.h>
+#include <Storages/TimeSeries/normalizeTimeSeriesDefinitionImpl.h>
 
 #include <Common/Exception.h>
 #include <Common/tests/gtest_global_register.h>
@@ -31,7 +31,6 @@ namespace DB::ErrorCodes
     extern const int BAD_TYPE_OF_FIELD;
     extern const int INCORRECT_QUERY;
     extern const int INVALID_SETTING_VALUE;
-    extern const int LOGICAL_ERROR;
     extern const int THERE_IS_NO_COLUMN;
 }
 
@@ -41,10 +40,7 @@ namespace
     {
         ParserCreateQuery parser;
         ASTPtr ast = parseQuery(parser, query, /* max_query_size = */ 0, DBMS_DEFAULT_MAX_PARSER_DEPTH, DBMS_DEFAULT_MAX_PARSER_BACKTRACKS);
-        auto * create_query = ast->as<ASTCreateQuery>();
-        if (!create_query)
-            throw std::runtime_error("Not a CREATE query: " + query);
-        return boost::intrusive_ptr<ASTCreateQuery>(create_query);
+        return boost::intrusive_ptr<ASTCreateQuery>(&ast->as<ASTCreateQuery &>());
     }
 
     ASTPtr parseExpression(const String & expression)
@@ -72,61 +68,40 @@ namespace
         return result;
     }
 
-    /// The columns of an external samples table.
-    ColumnsDescription externalSamplesColumns(const String & id_type = "UUID")
-    {
-        return makeColumns({makeColumn("id", id_type), makeColumn("timestamp", "DateTime64(3)"), makeColumn("value", "Float64")});
-    }
-
-    /// The columns of an external tags table.
-    ColumnsDescription externalTagsColumns(const String & id_type = "UUID", const String & id_default_expression = {})
-    {
-        return makeColumns({
-            makeColumn("id", id_type, id_default_expression),
-            makeColumn("metric_name", "LowCardinality(String)"),
-            makeColumn("tags", "Map(LowCardinality(String), String)"),
-            makeColumn("min_time", "Nullable(DateTime64(3))"),
-            makeColumn("max_time", "Nullable(DateTime64(3))")});
-    }
-
-    /// The columns of an external metrics table.
-    ColumnsDescription externalMetricsColumns()
-    {
-        return makeColumns({
-            makeColumn("metric_family_name", "String"),
-            makeColumn("type", "String"),
-            makeColumn("unit", "String"),
-            makeColumn("help", "String")});
-    }
-
     /// Normalizes the definition of a new table and returns the normalized definition as text.
-    String normalizeNewTable(const String & query, NormalizeTimeSeriesDefinitionInputs inputs = {})
+    String normalizeNewTable(const String & query, NormalizeTimeSeriesDefinitionParams params = {})
     {
         static const Settings default_query_settings;
-        if (!inputs.query_settings)
-            inputs.query_settings = &default_query_settings;
+        if (!params.query_settings)
+            params.query_settings = &default_query_settings;
         auto create_query = parseCreateQuery(query);
-        normalizeTimeSeriesDefinitionImpl(*create_query, LoadingStrictnessLevel::CREATE, /* is_restore_from_backup = */ false, inputs);
+        normalizeTimeSeriesDefinitionImpl(*create_query, params);
         return create_query->formatWithSecretsOneLine();
     }
 
-    /// Normalizes the definition of a new table and returns it as it would be stored,
-    /// e.g. to be used as the table from the clause `AS <other_table>`.
-    boost::intrusive_ptr<const ASTCreateQuery> storedDefinition(const String & query, NormalizeTimeSeriesDefinitionInputs inputs = {})
+    /// Normalizes the definition of a new table with the specified query-level settings and returns the normalized definition as text.
+    String normalizeNewTableWithSettings(const String & query, const Settings & query_settings)
     {
-        static const Settings default_query_settings;
-        if (!inputs.query_settings)
-            inputs.query_settings = &default_query_settings;
-        auto create_query = parseCreateQuery(query);
-        normalizeTimeSeriesDefinitionImpl(*create_query, LoadingStrictnessLevel::CREATE, /* is_restore_from_backup = */ false, inputs);
-        return create_query;
+        NormalizeTimeSeriesDefinitionParams params;
+        params.query_settings = &query_settings;
+        return normalizeNewTable(query, params);
+    }
+
+    /// Normalizes the definition of a new table with the clause `AS <old_table>` and returns the normalized definition as text.
+    /// `old_table_definition` is the stored definition of the old table, i.e. what the database catalog keeps for it.
+    String normalizeNewTableAs(const String & query, const String & old_table_definition, NormalizeTimeSeriesDefinitionParams params = {})
+    {
+        params.as_create_query = parseCreateQuery(old_table_definition);
+        return normalizeNewTable(query, params);
     }
 
     /// Normalizes the stored definition of an existing table (as on ATTACH) and returns the normalized definition as text.
     String normalizeExistingTable(const String & query)
     {
         auto create_query = parseCreateQuery(query);
-        normalizeTimeSeriesDefinitionImpl(*create_query, LoadingStrictnessLevel::ATTACH, /* is_restore_from_backup = */ false, /* inputs = */ {});
+        NormalizeTimeSeriesDefinitionParams params;
+        params.mode = LoadingStrictnessLevel::ATTACH;
+        normalizeTimeSeriesDefinitionImpl(*create_query, params);
         return create_query->formatWithSecretsOneLine();
     }
 
@@ -244,23 +219,37 @@ TEST_F(NormalizeTimeSeriesDefinitionTest, DefaultTableEngineChoosesInnerEngineFa
 {
     Settings query_settings;
     query_settings[Setting::default_table_engine] = DefaultTableEngine::ReplicatedMergeTree;
-    NormalizeTimeSeriesDefinitionInputs inputs;
-    inputs.query_settings = &query_settings;
 
-    auto definition = normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries", inputs);
+    auto definition = normalizeNewTableWithSettings("CREATE TABLE db.ts ENGINE = TimeSeries", query_settings);
     EXPECT_TRUE(extractInnerEngine(definition, "SAMPLES").starts_with("ReplicatedMergeTree ")) << definition;
     EXPECT_TRUE(extractInnerEngine(definition, "RECENT SAMPLES").starts_with("ReplicatedMergeTree ")) << definition;
     EXPECT_TRUE(extractInnerEngine(definition, "TAGS").starts_with("ReplicatedAggregatingMergeTree ")) << definition;
     EXPECT_TRUE(extractInnerEngine(definition, "METRICS").starts_with("ReplicatedReplacingMergeTree ")) << definition;
 
+    query_settings[Setting::default_table_engine] = DefaultTableEngine::SharedMergeTree;
+    definition = normalizeNewTableWithSettings("CREATE TABLE db.ts ENGINE = TimeSeries", query_settings);
+    EXPECT_TRUE(extractInnerEngine(definition, "SAMPLES").starts_with("SharedMergeTree ")) << definition;
+    EXPECT_TRUE(extractInnerEngine(definition, "RECENT SAMPLES").starts_with("SharedMergeTree ")) << definition;
+    EXPECT_TRUE(extractInnerEngine(definition, "TAGS").starts_with("SharedAggregatingMergeTree ")) << definition;
+    EXPECT_TRUE(extractInnerEngine(definition, "METRICS").starts_with("SharedReplacingMergeTree ")) << definition;
+
     /// A declared inner engine chooses the family of the other inner engines.
-    definition = normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries SAMPLES ENGINE = MergeTree", inputs);
+    definition = normalizeNewTableWithSettings("CREATE TABLE db.ts ENGINE = TimeSeries SAMPLES ENGINE = MergeTree", query_settings);
     EXPECT_TRUE(extractInnerEngine(definition, "SAMPLES").starts_with("MergeTree ")) << definition;
     EXPECT_TRUE(extractInnerEngine(definition, "TAGS").starts_with("AggregatingMergeTree ")) << definition;
 
+    query_settings[Setting::default_table_engine] = DefaultTableEngine::MergeTree;
+    definition = normalizeNewTableWithSettings("CREATE TABLE db.ts ENGINE = TimeSeries SAMPLES ENGINE = SharedMergeTree", query_settings);
+    EXPECT_TRUE(extractInnerEngine(definition, "SAMPLES").starts_with("SharedMergeTree ")) << definition;
+    EXPECT_TRUE(extractInnerEngine(definition, "TAGS").starts_with("SharedAggregatingMergeTree ")) << definition;
+
+    /// Only the MergeTree families can be the engines of the inner tables.
+    query_settings[Setting::default_table_engine] = DefaultTableEngine::Memory;
+    EXPECT_EQ(getExceptionCode([&] { normalizeNewTableWithSettings("CREATE TABLE db.ts ENGINE = TimeSeries", query_settings); }), ErrorCodes::INCORRECT_QUERY);
+
     /// Without a default engine the inner engines must be declared.
     query_settings[Setting::default_table_engine] = DefaultTableEngine::None;
-    EXPECT_EQ(getExceptionCode([&] { normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries", inputs); }), ErrorCodes::INCORRECT_QUERY);
+    EXPECT_EQ(getExceptionCode([&] { normalizeNewTableWithSettings("CREATE TABLE db.ts ENGINE = TimeSeries", query_settings); }), ErrorCodes::INCORRECT_QUERY);
 }
 
 
@@ -324,10 +313,20 @@ TEST_F(NormalizeTimeSeriesDefinitionTest, IdGeneratorSettingRecordsIdType)
 
 TEST_F(NormalizeTimeSeriesDefinitionTest, ExternalTagsTableRecordsIdTypeAndIdGenerator)
 {
-    NormalizeTimeSeriesDefinitionInputs inputs;
-    inputs.external_target_columns[ViewTarget::Tags] = externalTagsColumns("UInt64");
+    auto external_tags_columns = [](const String & id_type, const String & id_default_expression = {})
+    {
+        return makeColumns({
+            makeColumn("id", id_type, id_default_expression),
+            makeColumn("metric_name", "LowCardinality(String)"),
+            makeColumn("tags", "Map(LowCardinality(String), String)"),
+            makeColumn("min_time", "Nullable(DateTime64(3))"),
+            makeColumn("max_time", "Nullable(DateTime64(3))")});
+    };
 
-    auto definition = normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries TAGS db.ext_tags", inputs);
+    NormalizeTimeSeriesDefinitionParams params;
+    params.external_target_columns[ViewTarget::Tags] = external_tags_columns("UInt64");
+
+    auto definition = normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries TAGS db.ext_tags", params);
     EXPECT_TRUE(definition.contains("id_type = 'UInt64'")) << definition;
     EXPECT_TRUE(definition.contains("id_generator = 'sipHash64(tags)'")) << definition;
     EXPECT_TRUE(definition.contains("db.ext_tags")) << definition;
@@ -335,49 +334,61 @@ TEST_F(NormalizeTimeSeriesDefinitionTest, ExternalTagsTableRecordsIdTypeAndIdGen
     EXPECT_EQ(extractInnerColumns(definition, "SAMPLES"), "`id` UInt64, `timestamp` DateTime64(3) CODEC(DoubleDelta, ZSTD(1)), `value` Float64 CODEC(ZSTD(3))");
 
     /// The DEFAULT expression of the `id` column of the external table is recorded as the generator.
-    inputs.external_target_columns[ViewTarget::Tags] = externalTagsColumns("UInt64", "cityHash64(tags)");
-    definition = normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries TAGS db.ext_tags", inputs);
+    params.external_target_columns[ViewTarget::Tags] = external_tags_columns("UInt64", "cityHash64(tags)");
+    definition = normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries TAGS db.ext_tags", params);
     EXPECT_TRUE(definition.contains("id_generator = 'cityHash64(tags)'")) << definition;
 
     /// An explicit `id_generator` wins over the DEFAULT expression.
-    definition = normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries SETTINGS id_generator = 'sipHash64(tags)' TAGS db.ext_tags", inputs);
+    definition = normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries SETTINGS id_generator = 'sipHash64(tags)' TAGS db.ext_tags", params);
     EXPECT_TRUE(definition.contains("id_generator = 'sipHash64(tags)'")) << definition;
     EXPECT_FALSE(definition.contains("cityHash64")) << definition;
 
     /// The `id_type` setting must match the type of the `id` column of the external table.
     EXPECT_EQ(getExceptionCode([&]
     {
-        normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries SETTINGS id_type = 'UUID' TAGS db.ext_tags", inputs);
+        normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries SETTINGS id_type = 'UUID' TAGS db.ext_tags", params);
     }), ErrorCodes::BAD_TYPE_OF_FIELD);
 
     /// The external table must have all the required columns.
-    inputs.external_target_columns[ViewTarget::Tags] = makeColumns({makeColumn("id", "UInt64"), makeColumn("metric_name", "String")});
+    params.external_target_columns[ViewTarget::Tags] = makeColumns({makeColumn("id", "UInt64"), makeColumn("metric_name", "String")});
     EXPECT_EQ(getExceptionCode([&]
     {
-        normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries TAGS db.ext_tags", inputs);
+        normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries TAGS db.ext_tags", params);
     }), ErrorCodes::THERE_IS_NO_COLUMN);
 }
 
 
 TEST_F(NormalizeTimeSeriesDefinitionTest, ExternalTargetTablesDefineTypes)
 {
-    NormalizeTimeSeriesDefinitionInputs inputs;
-    inputs.external_target_columns[ViewTarget::Samples] = makeColumns(
+    /// The `min_time` and `max_time` columns must have the timestamp type of the external samples table, DateTime64(6) here.
+    auto external_tags_columns = [](const String & id_type)
+    {
+        return makeColumns({
+            makeColumn("id", id_type),
+            makeColumn("metric_name", "LowCardinality(String)"),
+            makeColumn("tags", "Map(LowCardinality(String), String)"),
+            makeColumn("min_time", "Nullable(DateTime64(6))"),
+            makeColumn("max_time", "Nullable(DateTime64(6))")});
+    };
+
+    NormalizeTimeSeriesDefinitionParams params;
+    params.external_target_columns[ViewTarget::Samples] = makeColumns(
         {makeColumn("id", "UInt64"), makeColumn("timestamp", "DateTime64(6)"), makeColumn("value", "Float32")});
-    inputs.external_target_columns[ViewTarget::Tags] = externalTagsColumns("UInt64");
-    inputs.external_target_columns[ViewTarget::Metrics] = externalMetricsColumns();
+    params.external_target_columns[ViewTarget::Tags] = external_tags_columns("UInt64");
+    params.external_target_columns[ViewTarget::Metrics] = makeColumns(
+        {makeColumn("metric_family_name", "String"), makeColumn("type", "String"), makeColumn("unit", "String"), makeColumn("help", "String")});
 
     auto definition = normalizeNewTable(
-        "CREATE TABLE db.ts ENGINE = TimeSeries SETTINGS recent_samples_ttl_seconds = 0 SAMPLES db.ext_samples TAGS db.ext_tags METRICS db.ext_metrics", inputs);
+        "CREATE TABLE db.ts ENGINE = TimeSeries SETTINGS recent_samples_ttl_seconds = 0 SAMPLES db.ext_samples TAGS db.ext_tags METRICS db.ext_metrics", params);
     EXPECT_TRUE(definition.contains("`time_series` Array(Tuple(DateTime64(6), Float32))")) << definition;
     EXPECT_TRUE(definition.contains("id_type = 'UInt64'")) << definition;
     EXPECT_FALSE(definition.contains("INNER")) << definition;
 
     /// The types of the external tables must match each other.
-    inputs.external_target_columns[ViewTarget::Tags] = externalTagsColumns("UUID");
+    params.external_target_columns[ViewTarget::Tags] = external_tags_columns("UUID");
     EXPECT_EQ(getExceptionCode([&]
     {
-        normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries SETTINGS recent_samples_ttl_seconds = 0 SAMPLES db.ext_samples TAGS db.ext_tags METRICS db.ext_metrics", inputs);
+        normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries SETTINGS recent_samples_ttl_seconds = 0 SAMPLES db.ext_samples TAGS db.ext_tags METRICS db.ext_metrics", params);
     }), ErrorCodes::BAD_TYPE_OF_FIELD);
 }
 
@@ -431,9 +442,8 @@ TEST_F(NormalizeTimeSeriesDefinitionTest, VersionSetting)
     EXPECT_EQ(getExceptionCode([] { normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries SETTINGS version = 999"); }), ErrorCodes::INVALID_SETTING_VALUE);
 
     /// The clause `AS <other_table>` doesn't copy the version: a new table gets the latest one.
-    NormalizeTimeSeriesDefinitionInputs inputs;
-    inputs.as_create_query = storedDefinition("CREATE TABLE db.src ENGINE = TimeSeries SETTINGS version = 0");
-    auto definition = normalizeNewTable("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries", inputs);
+    auto definition = normalizeNewTableAs("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries",
+        normalizeNewTable("CREATE TABLE db.src ENGINE = TimeSeries SETTINGS version = 0"));
     EXPECT_TRUE(definition.contains("version = 2")) << definition;
     EXPECT_FALSE(definition.contains("version = 0")) << definition;
 }
@@ -443,30 +453,29 @@ TEST_F(NormalizeTimeSeriesDefinitionTest, NormalizationIsIdempotent)
 {
     /// A stored definition is normalized again on ATTACH, and it can be replayed as a new table (e.g. on another replica):
     /// both must keep it unchanged, otherwise the generated parts would be applied twice.
-    NormalizeTimeSeriesDefinitionInputs inputs_with_external_tags;
-    inputs_with_external_tags.external_target_columns[ViewTarget::Tags] = externalTagsColumns("UInt64", "cityHash64(tags)");
+    NormalizeTimeSeriesDefinitionParams params_with_external_tags;
+    params_with_external_tags.external_target_columns[ViewTarget::Tags] = makeColumns({
+        makeColumn("id", "UInt64", "cityHash64(tags)"),
+        makeColumn("metric_name", "LowCardinality(String)"),
+        makeColumn("tags", "Map(LowCardinality(String), String)"),
+        makeColumn("min_time", "Nullable(DateTime64(3))"),
+        makeColumn("max_time", "Nullable(DateTime64(3))")});
 
-    const std::vector<std::pair<String, NormalizeTimeSeriesDefinitionInputs>> definitions =
+    const std::vector<std::pair<String, NormalizeTimeSeriesDefinitionParams>> definitions =
     {
         {"CREATE TABLE db.ts ENGINE = TimeSeries", {}},
         {"CREATE TABLE db.ts ENGINE = TimeSeries SETTINGS tags_to_columns = {'job': 'job'}, store_min_time_and_max_time = 0, recent_samples_ttl_seconds = 0", {}},
         {"CREATE TABLE db.ts ENGINE = TimeSeries SETTINGS recent_samples_partition_by = 'toStartOfHour(timestamp)' "
          "SAMPLES INNER COLUMNS (timestamp DateTime64(6) CODEC(Delta, ZSTD(1)), extra UInt8) TAGS INNER COLUMNS (id UInt64) "
          "TAGS ENGINE = AggregatingMergeTree ORDER BY (metric_name, id) SETTINGS index_granularity = 1024", {}},
-        {"CREATE TABLE db.ts ENGINE = TimeSeries TAGS db.ext_tags", inputs_with_external_tags},
+        {"CREATE TABLE db.ts ENGINE = TimeSeries TAGS db.ext_tags", params_with_external_tags},
     };
 
-    for (const auto & [query, inputs] : definitions)
+    for (const auto & [query, params] : definitions)
     {
-        auto stored = storedDefinition(query, inputs);
-        auto stored_text = stored->formatWithSecretsOneLine();
-
-        auto attached = boost::static_pointer_cast<ASTCreateQuery>(stored->clone());
-        normalizeTimeSeriesDefinitionImpl(*attached, LoadingStrictnessLevel::ATTACH, /* is_restore_from_backup = */ false, /* inputs = */ {});
-        EXPECT_EQ(attached->formatWithSecretsOneLine(), stored_text) << query;
-
-        auto replayed_text = normalizeNewTable(stored_text, inputs);
-        EXPECT_EQ(replayed_text, stored_text) << query;
+        auto stored_text = normalizeNewTable(query, params);
+        EXPECT_EQ(normalizeExistingTable(stored_text), stored_text) << query;
+        EXPECT_EQ(normalizeNewTable(stored_text, params), stored_text) << query;
     }
 }
 
@@ -475,12 +484,10 @@ TEST_F(NormalizeTimeSeriesDefinitionTest, CreateAsCopiesInnerDefinitions)
 {
     /// The clause `AS <other_table>` copies the inner columns and engines of the other table, except the parts generated
     /// for the other table: they are generated again for the new table from its settings, while the customized parts are copied.
-    NormalizeTimeSeriesDefinitionInputs inputs;
 
     /// A customized `min_time` column is not copied if the new table doesn't store `min_time`/`max_time`.
-    inputs.as_create_query = storedDefinition(
-        "CREATE TABLE db.src ENGINE = TimeSeries TAGS INNER COLUMNS (min_time SimpleAggregateFunction(min, Nullable(DateTime64(3))) CODEC(ZSTD(1)))");
-    auto definition = normalizeNewTable("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries SETTINGS store_min_time_and_max_time = 0", inputs);
+    auto definition = normalizeNewTableAs("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries SETTINGS store_min_time_and_max_time = 0",
+        normalizeNewTable("CREATE TABLE db.src ENGINE = TimeSeries TAGS INNER COLUMNS (min_time SimpleAggregateFunction(min, Nullable(DateTime64(3))) CODEC(ZSTD(1)))"));
     EXPECT_EQ(extractInnerColumns(definition, "TAGS"),
         "`id` " + default_id_type + " DEFAULT " + default_id_generator + ", `metric_name` LowCardinality(String), `tags` Map(LowCardinality(String), String)");
     EXPECT_EQ(extractInnerEngine(definition, "TAGS"),
@@ -488,9 +495,9 @@ TEST_F(NormalizeTimeSeriesDefinitionTest, CreateAsCopiesInnerDefinitions)
 
     /// `tags_to_columns` written in the query replaces the copied one: the column of a removed tag is not copied,
     /// the column of an added tag is generated.
-    inputs.as_create_query = storedDefinition("CREATE TABLE db.src ENGINE = TimeSeries SETTINGS tags_to_columns = {'job': 'job', 'instance': 'instance'}");
-    definition = normalizeNewTable(
-        "CREATE TABLE db.copy AS db.src ENGINE = TimeSeries SETTINGS tags_to_columns = {'instance': 'instance', 'region': 'region'}", inputs);
+    definition = normalizeNewTableAs(
+        "CREATE TABLE db.copy AS db.src ENGINE = TimeSeries SETTINGS tags_to_columns = {'instance': 'instance', 'region': 'region'}",
+        normalizeNewTable("CREATE TABLE db.src ENGINE = TimeSeries SETTINGS tags_to_columns = {'job': 'job', 'instance': 'instance'}"));
     EXPECT_EQ(extractInnerColumns(definition, "TAGS"),
         "`id` " + default_id_type + " DEFAULT " + default_id_generator + ", `metric_name` LowCardinality(String), `instance` String, `region` String, "
         "`tags` Map(LowCardinality(String), String), `min_time` SimpleAggregateFunction(min, Nullable(DateTime64(3))), "
@@ -498,13 +505,13 @@ TEST_F(NormalizeTimeSeriesDefinitionTest, CreateAsCopiesInnerDefinitions)
 
     /// The source customizes the codec of `timestamp` and adds an extra column in the samples table, sets the `id` type,
     /// and adds settings to the tags engine; everything else is generated.
-    inputs.as_create_query = storedDefinition(
+    const String src = normalizeNewTable(
         "CREATE TABLE db.src ENGINE = TimeSeries SAMPLES INNER COLUMNS (timestamp DateTime64(6) CODEC(Delta, ZSTD(1)), extra UInt8) "
         "TAGS INNER COLUMNS (id UInt64) TAGS ENGINE = AggregatingMergeTree ORDER BY (metric_name, id) SETTINGS index_granularity = 1024, min_bytes_for_wide_part = 0");
 
     /// The customized parts are copied, the generated parts follow the settings of the new table: `min_time`/`max_time`
     /// are not aggregated, so the tags engine is ReplacingMergeTree with them in the sorting key, and keeps its settings.
-    definition = normalizeNewTable("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries SETTINGS aggregate_min_time_and_max_time = 0", inputs);
+    definition = normalizeNewTableAs("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries SETTINGS aggregate_min_time_and_max_time = 0", src);
     EXPECT_EQ(extractInnerColumns(definition, "SAMPLES"),
         "`id` UInt64, `timestamp` DateTime64(6) CODEC(Delta, ZSTD(1)), `value` Float64 CODEC(ZSTD(3)), `extra` UInt8");
     EXPECT_EQ(extractInnerColumns(definition, "TAGS"),
@@ -516,7 +523,7 @@ TEST_F(NormalizeTimeSeriesDefinitionTest, CreateAsCopiesInnerDefinitions)
 
     /// A type declared in the query wins over the type of the other table, the other types are inherited: `value` is
     /// Float32 as declared, `timestamp` is DateTime64(6) as in the source. The declared samples columns replace the copied ones.
-    definition = normalizeNewTable("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries SAMPLES INNER COLUMNS (value Float32)", inputs);
+    definition = normalizeNewTableAs("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries SAMPLES INNER COLUMNS (value Float32)", src);
     EXPECT_EQ(extractInnerColumns(definition, "SAMPLES"), "`id` UInt64, `timestamp` DateTime64(6) CODEC(DoubleDelta, ZSTD(1)), `value` Float32");
     EXPECT_EQ(extractInnerColumns(definition, "RECENT SAMPLES"), "`id` UInt64, `timestamp` DateTime64(6) CODEC(DoubleDelta, ZSTD(1)), `value` Float32 CODEC(ZSTD(3))");
     EXPECT_EQ(extractInnerColumns(definition, "TAGS"),
@@ -525,10 +532,11 @@ TEST_F(NormalizeTimeSeriesDefinitionTest, CreateAsCopiesInnerDefinitions)
 
     /// The inner columns of the other table are not copied for a target replaced with an external table, the types come
     /// from the external table: `timestamp` is DateTime64(3) in the source and DateTime64(6) in the external table.
-    inputs.as_create_query = storedDefinition("CREATE TABLE db.src ENGINE = TimeSeries SAMPLES INNER COLUMNS (timestamp DateTime64(3) CODEC(ZSTD(5)))");
-    inputs.external_target_columns[ViewTarget::Samples] = makeColumns(
+    NormalizeTimeSeriesDefinitionParams params;
+    params.external_target_columns[ViewTarget::Samples] = makeColumns(
         {makeColumn("id", "UInt64"), makeColumn("timestamp", "DateTime64(6)"), makeColumn("value", "Float64")});
-    definition = normalizeNewTable("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries SAMPLES db.ext_data", inputs);
+    definition = normalizeNewTableAs("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries SAMPLES db.ext_data",
+        normalizeNewTable("CREATE TABLE db.src ENGINE = TimeSeries SAMPLES INNER COLUMNS (timestamp DateTime64(3) CODEC(ZSTD(5)))"), params);
     EXPECT_FALSE(definition.contains("ext_data SAMPLES INNER COLUMNS")) << definition;
     EXPECT_EQ(extractInnerColumns(definition, "RECENT SAMPLES"), "`id` UInt64, `timestamp` DateTime64(6) CODEC(DoubleDelta, ZSTD(1)), `value` Float64 CODEC(ZSTD(3))");
 }
@@ -539,8 +547,7 @@ TEST_F(NormalizeTimeSeriesDefinitionTest, CreateAsMergesSettings)
     /// The clause `AS <other_table>` copies the settings of the other table. A `SETTINGS` clause written in the query
     /// doesn't replace them: the settings are merged by name, so a written setting overrides the copied one, a written
     /// setting the other table doesn't have is added, and the rest of the other table's settings are still copied.
-    NormalizeTimeSeriesDefinitionInputs inputs;
-    inputs.as_create_query = storedDefinition(
+    const String src = normalizeNewTable(
         "CREATE TABLE db.src ENGINE = TimeSeries SETTINGS tags_to_columns = {'job': 'job'}, store_min_time_and_max_time = 0");
 
     const String tags_without_min_max = "`id` " + default_id_type + " DEFAULT " + default_id_generator
@@ -548,26 +555,26 @@ TEST_F(NormalizeTimeSeriesDefinitionTest, CreateAsMergesSettings)
 
     /// Without a `SETTINGS` clause: `job` comes from the copied `tags_to_columns`,
     /// and there is no `min_time`/`max_time` because the copied `store_min_time_and_max_time` is 0.
-    auto definition = normalizeNewTable("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries", inputs);
+    auto definition = normalizeNewTableAs("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries", src);
     EXPECT_EQ(extractInnerColumns(definition, "TAGS"), tags_without_min_max);
 
     /// The written `store_min_time_and_max_time` overrides the copied one so `min_time`/`max_time` appear,
     /// the written `tags_index_granularity` is added, and `tags_to_columns` is still copied.
-    definition = normalizeNewTable(
-        "CREATE TABLE db.copy AS db.src ENGINE = TimeSeries SETTINGS store_min_time_and_max_time = 1, tags_index_granularity = 4096", inputs);
+    definition = normalizeNewTableAs(
+        "CREATE TABLE db.copy AS db.src ENGINE = TimeSeries SETTINGS store_min_time_and_max_time = 1, tags_index_granularity = 4096", src);
     EXPECT_EQ(extractInnerColumns(definition, "TAGS"),
         tags_without_min_max + ", `min_time` SimpleAggregateFunction(min, Nullable(DateTime64(3))), `max_time` SimpleAggregateFunction(max, Nullable(DateTime64(3)))");
     EXPECT_TRUE(extractInnerEngine(definition, "TAGS").contains("index_granularity = 4096")) << definition;
 
     /// A setting written as `name = DEFAULT` is a mention of that setting too, so the value of the other table is not
     /// copied for it: the setting gets its default value and `min_time`/`max_time` appear, while `tags_to_columns` is still copied.
-    definition = normalizeNewTable(
-        "CREATE TABLE db.copy AS db.src ENGINE = TimeSeries SETTINGS aggregate_min_time_and_max_time = 0, store_min_time_and_max_time = DEFAULT", inputs);
+    definition = normalizeNewTableAs(
+        "CREATE TABLE db.copy AS db.src ENGINE = TimeSeries SETTINGS aggregate_min_time_and_max_time = 0, store_min_time_and_max_time = DEFAULT", src);
     EXPECT_EQ(extractInnerColumns(definition, "TAGS"), tags_without_min_max + ", `min_time` Nullable(DateTime64(3)), `max_time` Nullable(DateTime64(3))");
 
     /// A written value wins over a reset of the same setting, so there is no `min_time`/`max_time`.
-    definition = normalizeNewTable(
-        "CREATE TABLE db.copy AS db.src ENGINE = TimeSeries SETTINGS store_min_time_and_max_time = 0, store_min_time_and_max_time = DEFAULT", inputs);
+    definition = normalizeNewTableAs(
+        "CREATE TABLE db.copy AS db.src ENGINE = TimeSeries SETTINGS store_min_time_and_max_time = 0, store_min_time_and_max_time = DEFAULT", src);
     EXPECT_EQ(extractInnerColumns(definition, "TAGS"), tags_without_min_max);
 }
 
@@ -576,16 +583,14 @@ TEST_F(NormalizeTimeSeriesDefinitionTest, CreateAsTableWithExternalTargetTables)
 {
     /// The stored definition of a table with external target tables keeps the `id` type in the `id_type` setting,
     /// so the columns of the external tables are not needed to create a table `AS` that table.
-    NormalizeTimeSeriesDefinitionInputs inputs;
-    inputs.as_create_query = parseCreateQuery(
-        "CREATE TABLE db.src (`time_series` Array(Tuple(DateTime64(6), Float64))) ENGINE = TimeSeries "
+    const String src = "CREATE TABLE db.src (`time_series` Array(Tuple(DateTime64(6), Float64))) ENGINE = TimeSeries "
         "SETTINGS id_type = 'UInt64', id_generator = 'sipHash64(tags)', version = 2, recent_samples_ttl_seconds = 345600 "
-        "SAMPLES db.ext_samples RECENT SAMPLES db.ext_recent_samples TAGS db.ext_tags METRICS db.ext_metrics");
+        "SAMPLES db.ext_samples RECENT SAMPLES db.ext_recent_samples TAGS db.ext_tags METRICS db.ext_metrics";
 
-    auto definition = normalizeNewTable(
+    auto definition = normalizeNewTableAs(
         "CREATE TABLE db.copy AS db.src ENGINE = TimeSeries "
         "SAMPLES INNER COLUMNS (extra UInt8) RECENT SAMPLES INNER COLUMNS (extra UInt8) TAGS INNER COLUMNS (extra UInt8) METRICS INNER COLUMNS (extra UInt8)",
-        inputs);
+        src);
 
     EXPECT_TRUE(definition.contains("id_type = 'UInt64'")) << definition;
     EXPECT_TRUE(definition.contains("id_generator = 'sipHash64(tags)'")) << definition;
@@ -600,25 +605,23 @@ TEST_F(NormalizeTimeSeriesDefinitionTest, CreateAsTableWithExternalTargetTables)
     /// The external target tables of the other table are not copied, the new table must declare its own targets.
     EXPECT_EQ(getExceptionCode([&]
     {
-        normalizeNewTable("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries", inputs);
+        normalizeNewTableAs("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries", src);
     }), ErrorCodes::INCORRECT_QUERY);
 }
 
 
 TEST_F(NormalizeTimeSeriesDefinitionTest, CreateAsTableWithAnotherIdType)
 {
-    NormalizeTimeSeriesDefinitionInputs inputs;
-    inputs.as_create_query = parseCreateQuery(
-        "CREATE TABLE db.src ENGINE = TimeSeries SETTINGS id_type = 'UInt64', id_generator = 'sipHash64(tags)', version = 2, recent_samples_ttl_seconds = 345600 "
-        "SAMPLES INNER COLUMNS (`id` UInt64) TAGS INNER COLUMNS (`id` UInt64)");
+    const String src = "CREATE TABLE db.src ENGINE = TimeSeries SETTINGS id_type = 'UInt64', id_generator = 'sipHash64(tags)', version = 2, recent_samples_ttl_seconds = 345600 "
+        "SAMPLES INNER COLUMNS (`id` UInt64) TAGS INNER COLUMNS (`id` UInt64)";
 
     /// The settings written for the `id` type of the other table are copied only if the `id` type stays the same.
-    auto definition = normalizeNewTable("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries", inputs);
+    auto definition = normalizeNewTableAs("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries", src);
     EXPECT_TRUE(definition.contains("id_type = 'UInt64'")) << definition;
     EXPECT_TRUE(definition.contains("id_generator = 'sipHash64(tags)'")) << definition;
     EXPECT_TRUE(extractInnerColumns(definition, "TAGS").starts_with("`id` UInt64, ")) << definition;
 
-    definition = normalizeNewTable("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries TAGS INNER COLUMNS (id UUID)", inputs);
+    definition = normalizeNewTableAs("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries TAGS INNER COLUMNS (id UUID)", src);
     EXPECT_FALSE(definition.contains("id_type")) << definition;
     EXPECT_FALSE(definition.contains("id_generator")) << definition;
     EXPECT_TRUE(extractInnerColumns(definition, "TAGS").starts_with("`id` UUID DEFAULT reinterpretAsUUID(sipHash128(tags)), ")) << definition;
@@ -628,11 +631,16 @@ TEST_F(NormalizeTimeSeriesDefinitionTest, CreateAsTableWithAnotherIdType)
 
 TEST_F(NormalizeTimeSeriesDefinitionTest, EarlierVersionDoesNotRecordIdType)
 {
-    NormalizeTimeSeriesDefinitionInputs inputs;
-    inputs.external_target_columns[ViewTarget::Tags] = externalTagsColumns("UInt64");
+    NormalizeTimeSeriesDefinitionParams params;
+    params.external_target_columns[ViewTarget::Tags] = makeColumns({
+        makeColumn("id", "UInt64"),
+        makeColumn("metric_name", "LowCardinality(String)"),
+        makeColumn("tags", "Map(LowCardinality(String), String)"),
+        makeColumn("min_time", "Nullable(DateTime64(3))"),
+        makeColumn("max_time", "Nullable(DateTime64(3))")});
 
     /// A table pinned to version 1 is defined the way version 1 did it: an older server must be able to read it.
-    auto definition = normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries SETTINGS version = 1 TAGS db.ext_tags", inputs);
+    auto definition = normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries SETTINGS version = 1 TAGS db.ext_tags", params);
     EXPECT_TRUE(definition.contains("version = 1")) << definition;
     EXPECT_FALSE(definition.contains("id_type")) << definition;
     EXPECT_FALSE(definition.contains("id_generator")) << definition;
@@ -649,15 +657,12 @@ TEST_F(NormalizeTimeSeriesDefinitionTest, EarlierVersionDoesNotRecordIdType)
 
     /// The `id_type` copied from the other table is dropped if this table is pinned to an earlier version,
     /// while the `id` type is still inherited from it.
-    inputs.as_create_query = parseCreateQuery(
+    definition = normalizeNewTableAs(
+        "CREATE TABLE db.copy AS db.src ENGINE = TimeSeries SETTINGS version = 1 "
+        "SAMPLES INNER COLUMNS (extra UInt8) TAGS INNER COLUMNS (extra UInt8) METRICS INNER COLUMNS (extra UInt8)",
         "CREATE TABLE db.src (`time_series` Array(Tuple(DateTime64(3), Float64))) ENGINE = TimeSeries "
         "SETTINGS id_type = 'UInt64', id_generator = 'sipHash64(tags)', version = 2, recent_samples_ttl_seconds = 0 "
         "SAMPLES db.ext_samples TAGS db.ext_tags METRICS db.ext_metrics");
-    inputs.external_target_columns.clear();
-    definition = normalizeNewTable(
-        "CREATE TABLE db.copy AS db.src ENGINE = TimeSeries SETTINGS version = 1 "
-        "SAMPLES INNER COLUMNS (extra UInt8) TAGS INNER COLUMNS (extra UInt8) METRICS INNER COLUMNS (extra UInt8)",
-        inputs);
     EXPECT_TRUE(definition.contains("version = 1")) << definition;
     EXPECT_FALSE(definition.contains("id_type")) << definition;
     EXPECT_TRUE(definition.contains("id_generator = 'sipHash64(tags)'")) << definition;
@@ -668,16 +673,14 @@ TEST_F(NormalizeTimeSeriesDefinitionTest, EarlierVersionDoesNotRecordIdType)
 
 TEST_F(NormalizeTimeSeriesDefinitionTest, CreateAsNonTimeSeriesTable)
 {
-    NormalizeTimeSeriesDefinitionInputs inputs;
-    inputs.as_create_query = parseCreateQuery("CREATE TABLE db.src (x UInt8) ENGINE = MergeTree ORDER BY x");
-    EXPECT_EQ(getExceptionCode([&]
+    EXPECT_EQ(getExceptionCode([]
     {
-        normalizeNewTable("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries", inputs);
+        normalizeNewTableAs("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries", "CREATE TABLE db.src (x UInt8) ENGINE = MergeTree ORDER BY x");
     }), ErrorCodes::INCORRECT_QUERY);
 }
 
 
-TEST_F(NormalizeTimeSeriesDefinitionTest, ExistingTableNeedsNoInputs)
+TEST_F(NormalizeTimeSeriesDefinitionTest, ExistingTableNeedsNoParams)
 {
     /// On ATTACH the external target tables may not be loaded yet, so they are not read.
     auto definition = normalizeExistingTable(
@@ -714,27 +717,3 @@ TEST_F(NormalizeTimeSeriesDefinitionTest, ConvertsDefinitionsOfOlderVersions)
         normalizeNewTable("CREATE TABLE db.ts (`id` UInt64, `timestamp` DateTime64(3), `value` Float64) ENGINE = TimeSeries");
     }), ErrorCodes::INCORRECT_QUERY);
 }
-
-
-/// A logical error aborts the process in debug and sanitizer builds.
-#ifndef DEBUG_OR_SANITIZER_BUILD
-TEST_F(NormalizeTimeSeriesDefinitionTest, MissingInputsForNewTable)
-{
-    /// The wrapper `normalizeTimeSeriesDefinition` must pass everything a new table needs.
-    EXPECT_EQ(getExceptionCode([]
-    {
-        auto create_query = parseCreateQuery("CREATE TABLE db.ts ENGINE = TimeSeries");
-        normalizeTimeSeriesDefinitionImpl(*create_query, LoadingStrictnessLevel::CREATE, /* is_restore_from_backup = */ false, /* inputs = */ {});
-    }), ErrorCodes::LOGICAL_ERROR);
-
-    EXPECT_EQ(getExceptionCode([]
-    {
-        normalizeNewTable("CREATE TABLE db.ts ENGINE = TimeSeries TAGS db.ext_tags");
-    }), ErrorCodes::LOGICAL_ERROR);
-
-    EXPECT_EQ(getExceptionCode([]
-    {
-        normalizeNewTable("CREATE TABLE db.copy AS db.src ENGINE = TimeSeries");
-    }), ErrorCodes::LOGICAL_ERROR);
-}
-#endif
